requirements.txt
Flask
tensorflow
numpy
Pillow
scikit-learn
gunicorn
